+++ packages/moonstone/ExpandableInput/ExpandableInput.js
@@ -286,7 +286,6 @@ class ExpandableInputBase extends React.Component {
 				aria-label={this.calcAriaLabel()}
 				disabled={disabled}
 				label={this.calcLabel()}
-				noPointerMode
 				onClose={this.handleClose}
 				onMouseDown={this.handleMouseDown}
 				onSpotlightDisappear={onSpotlightDisappear}
@@ -336,6 +335,7 @@ class ExpandableInputBase extends React.Component {
  * @public
  */
 const ExpandableInput = Expandable(
+	{noPointerMode: true},
 	Changeable(
 		ExpandableInputBase
 	)