+++ packages/ui/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
-## Unreleased
+## [unreleased]
+
+### Added
+
+- `ui/Transition` prop `onShow` that fires when transitioning into view a component.
 
 ### Changed